Refactor overlapping UI components and eliminate 'AI slop' redundancies
- Created unified `EventSidebar` in `src/components/ui/` merging logic from `events` and `lab` features.
- Standardized `EventCard`, `SectionHeader`, `ContentCard`, and `GearCard` to use `Text` and `Box` primitives instead of raw Tailwind typography classes.
- Updated primary routes (`Dashboard`, `BlogDrafter`, `ResearchDetail`) to utilize the shared `PageHeader` component for consistent styling.
- Removed redundant feature-specific sidebar components.
- Verified changes with full project verification workflow including E2E and visual regression tests.
