+++ src/components/ui/GearCard.tsx
@@ -85,7 +85,7 @@ export function GearCard({
       />
       {verdict && (
         <Box display="flex" justify="end">
-          <Text variant="mono" size="xs" color="dim" className="text-text-body">
+          <Text variant="mono" size="xs" color="body">
             Best for: {verdict}
           </Text>
         </Box>
@@ -134,12 +134,14 @@ export function GearCard({
           variant="body"
           size="lg"
           weight="font-bold"
-          className="text-text-main leading-tight group-hover:text-accent transition-colors line-clamp-2"
+            color="main"
+            leading="tight"
+            className="group-hover:text-accent transition-colors line-clamp-2"
         >
           {title}
         </Text>
 
-        <Text variant="body" size="sm" color="dim" className="line-clamp-3 leading-relaxed text-text-body">
+        <Text variant="body" size="sm" color="dim" leading="relaxed" className="line-clamp-3">
            {excerpt}
         </Text>
       </Stack>
